+++ buildpacks/nodejs/CHANGELOG.md
@@ -4,6 +4,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 - Add license to buildpack.toml ([#17](https://github.com/heroku/buildpacks-node/pull/17))
+- Copy node modules directory path into the build ENV ([#15](https://github.com/heroku/buildpacks-node/pull/15))
 
 ## [0.7.1] 2021/01/20
 - Replace logging style to match style guides ([#63](https://github.com/heroku/nodejs-engine-buildpack/pull/63))